Add files via upload
Each of my four functions gave data for the MLB teams with the best pitching statistics during the 7th inning or later of a game. My data set includes all 30 teams in the MLB with their rank for the different categories that break down pitching. The different rows in my data set are the team’s Earned Run Average, total number of strikeouts, total number of earned runs and opponents’ batting average. 
My questions regarding each of my four functions:

What was the team with the lowest Earned Run Average? 
What was the team with the highest number of strikeouts? 
What was the median number of earned runs given up by each of the 30 teams? 
What was the team with the lowest opponents’ batting average?